scraper/requirements.txt
# Dependencias para los scrapers
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
jobspy>=0.4.1
undetected-chromedriver>=3.5.0
selenium>=4.15.0

# Dependencias para procesamiento y almacenamiento
openpyxl>=3.1.0
python-dotenv>=1.0.0

# Dependencias para Supabase e IA
supabase>=2.0.0
google-generativeai>=0.3.0
